flask/myflask/requirements.txt
about-time==3.1.1
aiohttp==3.8.1
aiosignal==1.2.0
alive-progress==2.4.1
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
attrdict==2.0.1
attrs==21.4.0
Babel==2.10.3
bce-python-sdk==0.8.74
blis==0.7.4
cachetools==5.2.0
captcha==0.4
catalogue==2.0.4
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.4
colorlog==6.7.0
commonmark==0.9.1
comtypes==1.1.14
cryptography==38.0.1
cssselect==1.1.0
cssutils==2.6.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.32
datasets==2.3.2
decorator==4.4.2
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.5.1
en-core-web-sm==3.0.0
et-xmlfile==1.1.0
executing==0.6.0
Faker==14.2.0
filelock==3.7.1
fire==0.4.0
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
fonttools==4.37.2
fpdf==1.7.2
frozenlist==1.3.0
fsspec==2022.8.2
future==0.18.2
grapheme==0.6.0
graphviz==0.16
greenlet==1.1.3
h5py==3.7.0
huggingface-hub==0.9.1
icecream==2.1.0
idna==2.10
imageio==2.22.0
imageio-ffmpeg==0.4.7
imgaug==0.4.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
jarowinkler==1.2.1
jieba==0.42.1
Jinja2==3.0.0
joblib==1.0.1
kiwisolver==1.3.1
libretranslatepy==2.1.1
lmdb==1.3.0
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
matplotlib==3.4.2
moviepy==1.0.3
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.5
networkx==2.5.1
nltk==3.6.2
numpy==1.20.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddlenlp==2.2.5
paddleocr==2.6.0.1
paddlepaddle==2.3.2
pandas==1.2.4
pathy==0.5.2
pdf2docx==0.5.6
pdfminer.six==20220524
pikepdf==5.6.0
Pillow==9.2.0
ply==3.11
poexcel==0.0.4
pofile==0.0.2
poimage==0.0.2
popdf==0.0.1
poppt==0.0.1
povideo==0.0.2
poword==0.0.1
premailer==3.10.0
preshed==3.0.5
proglog==0.1.10
protobuf==3.20.0
pyarrow==8.0.0
pyasn1==0.4.8
pyclipper==1.3.0.post3
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.7.4
Pygments==2.9.0
PyJWT==2.4.0
PyMuPDF==1.20.2
PyMySQL==1.0.2
PyOfficeRobot==0.0.10
pyparsing==2.4.7
PyPDF2==2.10.9
pypiwin32==223
pytextrank==3.1.1
python-dateutil==2.8.1
python-docx==0.8.11
python-office==0.2.2
pytz==2021.1
PyWavelets==1.3.0
pywifi==1.1.12
pywin32==304
PyYAML==6.0
qrcode==7.3.1
rapidfuzz==2.8.0
redis==4.3.4
regex==2021.4.4
reportlab==3.6.11
requests==2.25.1
responses==0.18.0
rich==12.5.1
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.9.1
search4file==0.1.6
seqeval==1.2.2
Shapely==1.8.4
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.4.41
srsly==2.4.1
termcolor==2.0.1
thinc==8.0.3
threadpoolctl==3.1.0
tifffile==2022.8.12
tqdm==4.64.1
translate==3.6.1
typer==0.3.2
typing_extensions==4.3.0
uiautomation==2.0.17
urllib3==1.26.4
visualdl==2.4.0
Wand==0.6.10
wasabi==0.8.2
Werkzeug==2.2.2
wftools==0.0.2
wordcloud==1.8.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
xxhash==3.0.0
yarl==1.7.2
you-get==0.4.1620
zipp==3.8.1
